+++ custom_components/houseplan/config_flow.py
@@ -2,7 +2,6 @@
 from __future__ import annotations
 
 import voluptuous as vol
-
 from homeassistant import config_entries
 
 from .const import CONF_ADMIN_ONLY, DOMAIN